+++ src/main/java/com/uaz/apirest/Titulo/TipoTitulacionRepository.java
@@ -2,5 +2,5 @@ package com.uaz.apirest.Titulo;
 
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
-public interface TipoTitulacionRepository extends Neo4jRepository<TipoTitulacion, Integer> {
+public interface TipoTitulacionRepository extends Neo4jRepository<TipoTitulacion, Long> {
 }
